+++ salesforce/lib/vf.py
@@ -6245,7 +6245,11 @@ tag_defs = {
                 "type": "String"
             },
             "value": {
-                "type": "Object"
+                "type": "Object",
+                "values": [
+                    "{!URLFOR(\$Resource.$1, '$2')}",
+                    "{!\$Resource.$1}"
+                ]
             }
         }
     },